refactor: adapters out 패키지로 이동
- 외부 API에서 데이터가 없는 경우 발생하는 예외로 adapters out에 위치하는 것이 적절하다 생각해 이동
