+++ src/main/java/com/srltas/runtogether/application/UserNeighborhoodService.java
@@ -4,12 +4,12 @@ import org.springframework.stereotype.Service;
 
 import com.srltas.runtogether.application.port.in.AddUserNeighborhood;
 import com.srltas.runtogether.application.port.in.AddUserNeighborhoodCommand;
-import com.srltas.runtogether.domain.model.neighborhood.exception.NeighborhoodNotFoundException;
+import com.srltas.runtogether.adapter.out.exception.NeighborhoodNotFoundException;
 import com.srltas.runtogether.domain.model.neighborhood.Neighborhood;
 import com.srltas.runtogether.domain.model.neighborhood.NeighborhoodRepository;
 import com.srltas.runtogether.domain.model.user.User;
 import com.srltas.runtogether.domain.model.user.UserRepository;
-import com.srltas.runtogether.domain.model.user.exception.UserNotFoundException;
+import com.srltas.runtogether.adapter.out.exception.UserNotFoundException;
 
 import lombok.RequiredArgsConstructor;
