+++ src/main/java/com/srltas/runtogether/common/exception/handler/RunTogetherExceptionHandler.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.resource.NoResourceFoundException;
 
 import com.srltas.runtogether.common.exception.BusinessException;
-import com.srltas.runtogether.common.exception.EntityNotFoundException;
+import com.srltas.runtogether.adapter.out.exception.EntityNotFoundException;
 import com.srltas.runtogether.common.exception.UnauthorizedException;
 
 @RestControllerAdvice
